# swagger.yaml
openapi: 3.0.0
info:
  title: SquidFlow Platform
  version: '1.0.0'
  description: API for managing applications, clusters, templates and secrets
...
servers:
  - url: http://localhost:38080/api/v1
    description: Local development server
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
    ApplicationSource:
      type: object
      required:
        - type
        - url
        - targetRevision
        - path
      properties:
        type:
          type: string
          enum: [git]
          example: "git"
        url:
          type: string
          example: "git@github.com:squidflow/platform.git"
        targetRevision:
          type: string
          example: "main"
        path:
          type: string
          example: "manifest/fluent-operator"
...
    TLS:
      type: object
      properties:
        enabled:
          type: boolean
          example: true
        secretName:
          type: string
          example: "demo1-tls"
    Ingress:
      type: object
      properties:
        host:
          type: string
          example: "demo1.h4.com"
        tls:
          $ref: '#/components/schemas/TLS'
    SecretStoreRef:
      type: object
      properties:
        id:
          type: string
          example: "secret-store"
    ExternalSecret:
      type: object
      properties:
        secret_store_ref:
          $ref: '#/components/schemas/SecretStoreRef'
    Security:
      type: object
      properties:
        external_secret:
          $ref: '#/components/schemas/ExternalSecret'
...
    ApplicationResponse:
      type: object
      properties:
        message:
          type: string
          example: "Applications created successfully"
        application:
          $ref: '#/components/schemas/ApplicationDetail'
    ApplicationDetail:
      type: object
      properties:
        name:
          type: string
        tenant_name:
          type: string
        appcode:
          type: string
        description:
          type: string
        created_by:
          type: string
        template:
          type: object
          properties:
            source:
              $ref: '#/components/schemas/ApplicationSource'
            last_commit_info:
              type: object
              properties:
                commit_id:
                  type: string
                message:
                  type: string
                author:
                  type: string
                timestamp:
                  type: string
                  format: date-time
        runtime_status:
          type: object
          properties:
            status:
              type: string
            health:
              type: string
            sync_status:
              type: string
            deployed_clusters:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                    example: "cluster-1"
                  namespace:
                    type: string
                    example: "default"
                  pod_count:
                    type: integer
                    example: 3
                  secret_count:
                    type: integer
                    example: 2
                  status:
                    type: string
                    example: "Running"
                  last_sync_time:
                    type: string
                    format: date-time
            resource_metrics:
              $ref: '#/components/schemas/ResourceMetrics'
    ResourceMetrics:
      type: object
      properties:
        cpu_cores:
          type: string
          example: "200m"
        memory_usage:
          type: string
          example: "256Mi"
...
paths:
  /deploy/applications:
    post:
      tags: [Application]
      summary: Create application
      operationId: ApplicationCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - application_source
                - application_instantiation
              properties:
                application_source:
                  type: object
                  required:
                    - repo
                    - target_revision
                    - path
                  properties:
                    repo:
                      type: string
                    target_revision:
                      type: string
                    path:
                      type: string
                    submodules:
                      type: boolean
                application_instantiation:
                  type: object
                  required:
                    - application_name
                    - tenant_name
                  properties:
                    application_name:
                      type: string
                    tenant_name:
                      type: string
                    appcode:
                      type: string
                    description:
                      type: string
                application_target:
                  type: array
                  items:
                    type: object
                    required:
                      - cluster
                      - namespace
                    properties:
                      cluster:
                        type: string
                      namespace:
                        type: string
                is_dryrun:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'

    get:
      tags: [Application]
      summary: List applications
      operationId: ApplicationsList
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApplicationDetail'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deploy/applications/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Application name

    get:
      tags: [Application]
      summary: Get application details
      description: Get detailed information about an application including template source, H4 configuration and runtime status
      operationId: ApplicationGet
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetail'
        '403':
          description: Forbidden - user does not have permission to view this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Application]
      summary: Delete application
      description: Delete an application. If this is the last project using the application, the entire application directory will be removed.
      operationId: ApplicationDelete
      responses:
        '204':
          description: Application deleted successfully
        '403':
          description: Forbidden - user does not have permission to delete this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                gitError:
                  value:
                    error: "Failed to push changes to git repository"
                argocdError:
                  value:
                    error: "Failed to delete ArgoCD application"

    patch:
      tags: [Application]
      summary: Update application
      operationId: ApplicationUpdate
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Application name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                application_source:
                  type: object
                  properties:
                    repo:
                      type: string
                    target_revision:
                      type: string
                    path:
                      type: string
                    submodules:
                      type: boolean
                application_instantiation:
                  type: object
                  properties:
                    description:
                      type: string
                    security:
                      $ref: '#/components/schemas/Security'
                    ingress:
                      $ref: '#/components/schemas/Ingress'
      responses:
        '200':
          description: Application updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Application updated successfully"
                  application:
                    $ref: '#/components/schemas/ApplicationDetail'
        '403':
          description: Forbidden - user does not have permission to update this application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /deploy/applications/validate:
    post:
      tags: [Application]
      summary: Validate application source
      operationId: ApplicationSourceValidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repo:
                  type: string
                  example: "git@github.com:SquidFlow/platform.git"
                target_revision:
                  type: string
                  example: "main"
                path:
                  type: string
                  example: "overlays/app/fluent-operator"
                submodules:
                  type: boolean
                  example: true
                application_specifier:
                  type: object
                  properties:
                    helm_manifest_path:
                      type: string
                      example: "manifests/4.0.0"
      responses:
        '200':
          description: Application source validated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Valid kustomize application source"
                  type:
                    type: string
                    example: "kustomize"
                  suiteableEnv:
                    type: array
                    items:
                      type: object
                      properties:
                        environments:
                          type: string
                          example: "sit"
                        valid:
                          type: boolean
                          example: true
                        error:
                          type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...